Update 03.auto-train-models.ipynb
Basically current local_run.register_model() fails as it does not specify which iteration to register. 
To fix this, first, I've set the exit_score to 0.9995. This allows we see all 20 interations.
Then, I added description that iteration=9 is the best model, confirmed from Widget.
Lastly, I've added iteration=0 as a parameter to register_model() call. This ensures successful registration of the model.